Create MovieCard.css
Movie Card design created so that for right now it appears purple and has a hover effect so that the cards grow in size when the user hovers over them. As of right now just basic designs used from the playlist website, will update to fit new theme with time.