Create MovieModal.css
Movie Modal css created so that when the modal is displayed to first darken the background, appear card like as if it were floating above the screen. Also, some cool features have been added so that the modal fades in when the movie is clicked. You can exit out through either clicking the exit button on the modal or outside of the modal